Move autodiscovery function to dedicated go file
